Rename RcbUi and RcbInner to PurchasesUi and PurchasesUiInner
