+++ src/stories/utils/get-story-checkout-args.ts
@@ -1,9 +1,9 @@
 import { type ComponentProps } from "svelte";
-import type RcbUiInner from "src/ui/rcb-ui-inner.svelte";
+import type PurchasesInner from "../../../ui/purchases-ui-inner.svelte";
 import { buildCheckoutStartResponse } from "./purchase-response-builder";
 
 export const getArgs = async (): Promise<
-  Partial<ComponentProps<RcbUiInner>>
+  Partial<ComponentProps<PurchasesInner>>
 > => {
   return {
     paymentInfoCollectionMetadata: await buildCheckoutStartResponse(),
